Process to create a MODS files from a single EAD file
Early draft of a process to creates MODS files from a single EAD file.
The output is one MODS file for every level of description in the EAD
file, but one extra MODS file that contains a simplifed "context tree"
of the entire collection hierarchy that's derived from the EAD file.
Each MODS file also retains an EAD snippet of the level of description
from which it was generated.